update assimp to d228dbeab7b6f2ee71a0a5744ddee2d87851fe11
basicly this is a sed -i 's/float/double/g' $(find -type f)

THIS VERY LIKELY WILL CAUSE DESYNCS / CRASHES / YOUR COMPUTER EXPLODE!

you have been warned! ;)

+++ rts/lib/assimp/code/BlenderDNA.cpp
@@ -330,9 +330,9 @@ void DNA :: AddPrimitiveStructures()
 	structures.back().size = 4;
 
 
-	indices["double"] = structures.size();
+	indices["float"] = structures.size();
 	structures.push_back( Structure() );
-	structures.back().name = "double";
+	structures.back().name = "float";
 	structures.back().size = 8;
 
 	// no long, seemingly.
